+++ src/bash_internal.h
@@ -27,6 +27,7 @@ typedef struct {
     int function_depth;
     int handler_depth;
     bool has_cleanup_helpers;
+    bool has_signal_handlers;
     size_t temp_counter;
     size_t handler_counter;
     bool needs_case_types;
@@ -81,6 +82,7 @@ bool program_uses_case(const DsLowerProgram *program);
 bool program_uses_int_helpers(const DsLowerProgram *program);
 bool program_uses_function_value_helpers(const DsLowerProgram *program);
 bool program_uses_handlers(const DsLowerProgram *program);
+bool program_uses_signal_handlers(const DsLowerProgram *program);
 
 bool emit_block_body(BashEmitter *e, const DsLowerStmt *block, int indent);
 bool emit_function(BashEmitter *e, const DsLowerFn *fn);